+++ packages/types/src/global.ts
@@ -10,6 +10,7 @@ type ApiAt = ApiDecoration<'promise'> & {rpc: ApiPromise['rpc']};
 
 declare global {
   const api: ApiAt;
+  const unsafeApi: ApiPromise;
   const logger: Pino.Logger;
   const store: Store;
   const chainId: string;